requirements.txt
streamlit==1.44.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.1
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==2.2.4
